+++ app/src/main/res/drawable/selector_tablayout.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <!-- 누를 때 상태 -->
+    <item android:state_selected="true">
+        <shape>
+            <solid android:color="#E0E0E0" />
+            <corners android:radius="10dp" />
+        </shape>
+    </item>
+
+    <!-- 기본 상태 (투명한 배경) -->
+    <item>
+        <shape>
+            <solid android:color="@android:color/transparent" />
+            <corners android:radius="10dp" />
+        </shape>
+    </item>
+</selector>
